Fix printf warnings
- GCC wants to see "long long".

+++ SurgSim/Testing/MlcpIO/ReadText.cpp
@@ -268,7 +268,8 @@ static bool readEigenMatrix(const std::string& fileName, FILE* in, const char* l
 			if (rowVector.cols() != numCols)
 			{
 				fprintf(stderr, "Inconsistent number of columns for Eigen matrix (%lld vs %lld)\n  in file '%s'\n",
-						static_cast<int64_t>(numCols), static_cast<int64_t>(rowVector.cols()), fileName.c_str());
+						static_cast<long long int>(numCols), static_cast<long long int>(rowVector.cols()),
+						fileName.c_str());
 				return false;
 			}
 			matrix->conservativeResize(newNumRows, numCols);
